Rename directory to motion_control
